+++ layers/lang/rust.lua
@@ -3,34 +3,6 @@ local lsp = require("LYRD.layers.lsp")
 
 local L = { name = "Rust Language" }
 
-local function setup_dap()
-	local dap = require("dap")
-	dap.adapters.codelldb = require("LYRD.shared.dap.codelldb")
-	local debug_configuration = {
-		{
-			name = "runit",
-			type = "codelldb",
-			request = "launch",
-
-			program = function()
-				return vim.fn.input("", vim.fn.getcwd(), "file")
-			end,
-
-			args = { "--log_level=all" },
-			cwd = "${workspaceFolder}",
-			stopOnEntry = false,
-			terminal = "integrated",
-
-			pid = function()
-				local handle = io.popen("pgrep hw$")
-				local result = handle:read()
-				handle:close()
-				return result
-			end,
-		},
-	}
-	dap.configurations.rust = debug_configuration
-end
 function L.plugins()
 	setup.plugin({
 		{
@@ -52,7 +24,8 @@ function L.preparation()
 end
 
 function L.settings()
-	setup_dap()
+	local debugger = require("LYRD.shared.dap.codelldb")
+	debugger.setup({ "rust" })
 end
 
 function L.complete() end
